changed doubles to float to be more cost effective

+++ src/Header.hpp
@@ -2,20 +2,20 @@
 #ifndef HEADER_HPP
 #define HEADER_HPP
 class PositionSensing{
-    public: PositionSensing(double startingX, double startingY, double leftTrackingWheelDistance, double rightTrackingWheelDistance, double backTrackingWheelDistance, double startingOrientation);
-    public: std::vector<double> GetPosition();
-    public: void UpdatePosition(double deltaLeft, double deltaRight, double deltaBack);
+    public: PositionSensing(float startingX, float startingY, float leftTrackingWheelDistance, float rightTrackingWheelDistance, float backTrackingWheelDistance, float startingOrientation);
+    public: std::vector<float> GetPosition();
+    public: void UpdatePosition(float deltaLeft, float deltaRight, float deltaBack);
 };
 
 class VectorMath {
-    public: static void CartesianToPolar(std::vector<double> componentVector);
-    public: static void PolarToCartesian(std::vector<double> polarVector);
-    public: static double AngleToDistance(double angle);
-    public: static double RadiansToDegrees(double radians);
-    public: static double DegreesToRadians(double degrees);
-    public: static void ScaleVector(std::vector<double> &cartesianVector, double scalar);
-    public: static void AdditionVector(std::vector<double> &Vector, double constant);
-    public: static void RotateVectorAddition(std::vector<double> &polarVector, double radians);
-    public: static void AddVectors(std::vector<double> &holderVector, std::vector<double> &addedVector);
+    public: static void CartesianToPolar(std::vector<float> componentVector);
+    public: static void PolarToCartesian(std::vector<float> polarVector);
+    public: static float AngleToDistance(float angle);
+    public: static float RadiansToDegrees(float radians);
+    public: static float DegreesToRadians(float degrees);
+    public: static void ScaleVector(std::vector<float> &cartesianVector, float scalar);
+    public: static void AdditionVector(std::vector<float> &Vector, float constant);
+    public: static void RotateVectorAddition(std::vector<float> &polarVector, float radians);
+    public: static void AddVectors(std::vector<float> &holderVector, std::vector<float> &addedVector);
 };
 #endif
